Renaming and fixing tests
Byte it smoke tests
