+++ src/test/kotlin/coverosR3z/timerecording/EnteringTimeBDD.kt
@@ -66,48 +66,6 @@ class EnteringTimeBDD {
         assertThrows(ExceededDailyHoursAmountException::class.java) { tru.recordTime(entry) }
     }
 
-
-
-    @Test
-    fun `API_TEST An employee should be able to enter time for previous day`() {
-        // given the employee worked 8 hours yesterday
-        val pmd = PureMemoryDatabase()
-        val authPersistence = AuthenticationPersistence(pmd)
-        val au = AuthenticationUtilities(authPersistence)
-
-        val systemTru = TimeRecordingUtilities(TimeEntryPersistence(pmd), CurrentUser(SYSTEM_USER))
-        val alice = systemTru.createEmployee(EmployeeName("Alice"))
-        val userName = UserName("alice_1")
-
-        au.register(userName, DEFAULT_PASSWORD, alice.id)
-        val (_, user) = au.login(userName, DEFAULT_PASSWORD)
-
-        val newProject = systemTru.createProject(DEFAULT_PROJECT_NAME)
-
-        val tru = TimeRecordingUtilities(TimeEntryPersistence(pmd), CurrentUser(user))
-
-        assertTrue("Registration must have succeeded", au.isUserRegistered(userName))
-
-        val eightHours = 8*60
-        val entry = createTimeEntryPreDatabase(time = Time(eightHours),
-            project = newProject,
-            employee = alice)
-
-        val descr = "no problem here"
-        // when the employee enters their time
-        val data: Map<String, String> = mapOf(EnterTimeAPI.Elements.PROJECT_INPUT.elemName to entry.project.id.value.toString(),
-                         EnterTimeAPI.Elements.TIME_INPUT.elemName to entry.time.numberOfMinutes.toString(),
-                         EnterTimeAPI.Elements.DATE_INPUT.elemName to A_RANDOM_DAY_IN_JUNE_2020.stringValue,
-                         EnterTimeAPI.Elements.DETAIL_INPUT.elemName to descr)
-        val result = EnterTimeAPI.handlePOST(tru, user.employeeId, data = data)
-
-        // then time is saved
-        val message = "we should have gotten the success page.Got: $result"
-        assertTrue(message, toStr(result.fileContents).contains("SUCCESS"))
-        assertTrue(tru.getEntriesForEmployeeOnDate(alice.id, A_RANDOM_DAY_IN_JUNE_2020).any {
-            it.details.value == descr })
-    }
-
     @Test
     fun `An employee should be able to edit time from a previous time entry` () {
         //given Andrea has a previous time entry
